reorg zmq and ws
