+++ bokeh/server/websocket/__init__.py
@@ -1,0 +1,3 @@
+
+from .handler import WebSocketHandler
+from .manager import WebSocketManager
